refactor: DDD기반 구조로 변경
